Add addDebtScreen and function to add into database

+++ app/src/main/java/com/angelpr/wallet/data/db/entities/DebtsWalletEntity.kt
@@ -13,7 +13,7 @@ data class DebtsWalletEntity(
     @ColumnInfo(name = "idWallet")
     val idWallet: Int,
     @ColumnInfo(name = "debt")
-    val debt: Int,
+    val debt: Float,
     @ColumnInfo(name = "type")
     val type: String,
     @ColumnInfo(name = "isPaid")
