+++ app/src/main/java/com/angelpr/wallet/data/model/DebtModel.kt
@@ -5,7 +5,7 @@ import androidx.room.ColumnInfo
 data class DebtModel(
     val id: Int = 0,
     val idWallet: Int,
-    val debt: Int,
+    val debt: Float,
     val type: String,
     val isPaid: Int,
     val date: Long
